Cleaned up the design and API for KidnummerValidaotr and Kidnummer
Added more XMLdoc
Made the constructor for Kidnummer protected internal as to avoid
confusion.
The only way to create an instance of Kidnummer is to use the
GetKidnummer()method in KidnummerValidator. This ensures that the user
gets a validated instance back.
This might be improved as it conflicts with SRP (both validator and
factory)